+++ src/main/java/com/group4/app/controller/CreatureStatController.java
@@ -1,9 +1,9 @@
 package com.group4.app.controller;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
 import com.group4.app.model.creatures.IAttackable;
-import com.group4.app.model.creatures.IPositionable;
+import com.group4.app.model.dungeon.IPositionable;
+import com.group4.app.model.dungeon.Position;
 
 public class CreatureStatController {
     public int getCreatureHealth(Position creaturePosition) {
