+++ src/main/java/com/group4/app/controller/InventoryController.java
@@ -2,8 +2,8 @@ package com.group4.app.controller;
 
 import com.group4.app.model.Model;
 import com.group4.app.model.actions.Action;
-import com.group4.app.model.actions.ItemActionInput;
 import com.group4.app.model.creatures.Player;
+import com.group4.app.model.input.ItemActionInput;
 
 public class InventoryController {
     public void useItem(String itemName) {
