+++ src/main/java/com/group4/app/controller/worldControllers/PlayerMovementController.java
@@ -5,8 +5,8 @@ import java.util.Set;
 
 import com.group4.app.controller.ActionController;
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
-import com.group4.app.model.actions.PositionActionInput;
+import com.group4.app.model.dungeon.Position;
+import com.group4.app.model.input.PositionActionInput;
 import com.group4.app.view.ActionState;
