+++ src/main/java/com/group4/app/model/IController.java
@@ -1,6 +1,6 @@
 package com.group4.app.model;
 
-import com.group4.app.model.actions.ActionInput;
+import com.group4.app.model.input.ActionInput;
 
 public interface IController {
     ActionInput<?> getActionInput();
