--- src/main/java/com/group4/app/model/IDrawable.java
@@ -1,7 +1,0 @@
-package com.group4.app.model;
-
-import com.group4.app.model.creatures.IPositionable;
-
-public interface IDrawable extends IPositionable{
-    String getId();
-}
