+++ src/main/java/com/group4/app/model/Model.java
@@ -8,20 +8,25 @@ import java.util.Map;
 import java.util.Set;
 
 import com.group4.app.controller.ActionController;
-import com.group4.app.model.actions.ActionInput;
 import com.group4.app.model.creatures.AttributeType;
 import com.group4.app.model.creatures.Enemy;
 import com.group4.app.model.creatures.EnemyFactory;
 import com.group4.app.model.creatures.Entity;
 import com.group4.app.model.creatures.IAttackable;
-import com.group4.app.model.creatures.IPositionable;
 import com.group4.app.model.creatures.Player;
 import com.group4.app.model.dungeon.DungeonEntitySpawner;
 import com.group4.app.model.dungeon.DungeonWorldGenerator;
+import com.group4.app.model.dungeon.IDrawable;
+import com.group4.app.model.dungeon.IPositionable;
 import com.group4.app.model.dungeon.IWorldContainer;
+import com.group4.app.model.dungeon.PathfindingHelper;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.dungeon.Tile;
 import com.group4.app.model.dungeon.World;
+import com.group4.app.model.input.ActionInput;
 import com.group4.app.model.items.WeaponFactory;
+import com.group4.app.model.turns.ITurnTaker;
+import com.group4.app.model.turns.TurnHandler;
 
 public class Model implements IWorldContainer {
     private static Model instance = null;
