+++ src/main/java/com/group4/app/model/actions/IAction.java
@@ -2,8 +2,8 @@ package com.group4.app.model.actions;
 
 import java.util.Set;
 
-import com.group4.app.model.Position;
-import com.group4.app.model.creatures.IPositionable;
+import com.group4.app.model.dungeon.IPositionable;
+import com.group4.app.model.dungeon.Position;
 
 public interface IAction<T> {
     void perform(T target);
