+++ src/main/java/com/group4/app/model/actions/IActionTaker.java
@@ -2,7 +2,7 @@ package com.group4.app.model.actions;
 
 import com.group4.app.model.creatures.IAttackable;
 import com.group4.app.model.creatures.ICanAttack;
-import com.group4.app.model.creatures.IPositionable;
+import com.group4.app.model.dungeon.IPositionable;
 
 public interface IActionTaker extends IPositionable{
