+++ src/main/java/com/group4/app/model/actions/ItemAction.java
@@ -1,7 +1,7 @@
 package com.group4.app.model.actions;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.items.IInventoriable;
 import com.group4.app.model.items.IUser;
 import com.group4.app.model.items.Inventory;
