+++ src/main/java/com/group4/app/model/creatures/EnemyFactory.java
@@ -1,6 +1,6 @@
 package com.group4.app.model.creatures;
 
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.items.WeaponFactory;
 
 public class EnemyFactory {
