+++ src/main/java/com/group4/app/model/creatures/ICanAttack.java
@@ -1,6 +1,6 @@
 package com.group4.app.model.creatures;
 
-import com.group4.app.model.IHasPosition;
+import com.group4.app.model.dungeon.IHasPosition;
 
 public interface ICanAttack extends IHasPosition {
     int getDamage();
