+++ src/main/java/com/group4/app/model/dungeon/IDrawable.java
@@ -1,0 +1,5 @@
+package com.group4.app.model.dungeon;
+
+public interface IDrawable extends IPositionable{
+    String getId();
+}
