+++ src/main/java/com/group4/app/model/dungeon/ITileContainer.java
@@ -1,7 +1,5 @@
 package com.group4.app.model.dungeon;
 
-import com.group4.app.model.Position;
-
 public interface ITileContainer {
     void add(Tile tile);
     void remove(Tile tile);
