+++ src/main/java/com/group4/app/model/dungeon/Tile.java
@@ -3,10 +3,7 @@ package com.group4.app.model.dungeon;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.group4.app.model.IDrawable;
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
-import com.group4.app.model.creatures.IPositionable;
 
 
 public class Tile implements IDrawable {
