+++ src/main/java/com/group4/app/model/input/AttackActionInput.java
@@ -1,4 +1,4 @@
-package com.group4.app.model.actions;
+package com.group4.app.model.input;
 
 import com.group4.app.model.creatures.IAttackable;
