+++ src/main/java/com/group4/app/model/input/EndTurnActionInput.java
@@ -1,7 +1,7 @@
-package com.group4.app.model.actions;
+package com.group4.app.model.input;
 
-import com.group4.app.model.ITurnTaker;
 import com.group4.app.model.creatures.Creature;
+import com.group4.app.model.turns.ITurnTaker;
 
 
 public class EndTurnActionInput extends ActionInput<Creature>{
