+++ src/main/java/com/group4/app/model/input/ItemActionInput.java
@@ -1,4 +1,4 @@
-package com.group4.app.model.actions;
+package com.group4.app.model.input;
 
 import com.group4.app.model.items.IInventoriable;
