+++ src/main/java/com/group4/app/model/turns/ITurnTaker.java
@@ -1,4 +1,4 @@
-package com.group4.app.model;
+package com.group4.app.model.turns;
 
 public interface ITurnTaker {
     void takeTurn();
