+++ src/main/java/com/group4/app/view/EnemyHealthBar.java
@@ -7,7 +7,7 @@ import javax.swing.JPanel;
 import javax.swing.JProgressBar;
 
 import com.group4.app.controller.CreatureStatController;
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.Position;
 
 /**
  * A health bar for a creature at the given position.
