+++ src/main/java/com/group4/app/view/worldView/WorldView.java
@@ -17,8 +17,8 @@ import com.group4.app.controller.worldControllers.AWorldController;
 import com.group4.app.controller.worldControllers.PlayerMovementController;
 import com.group4.app.controller.worldControllers.PlayerViewAttackController;
 import com.group4.app.controller.PositionController;
-import com.group4.app.model.IDrawable;
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.IDrawable;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.view.ActionState;
 import com.group4.app.view.EnemyHealthBar;
 import com.group4.app.view.SubView;
