+++ src/test/java/com/group4/app/testmodel/TestAttackAction.java
@@ -7,7 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
 import com.group4.app.model.actions.Action;
 import com.group4.app.model.actions.AttackAction;
 import com.group4.app.model.creatures.Enemy;
@@ -15,6 +14,7 @@ import com.group4.app.model.creatures.EnemyFactory;
 import com.group4.app.model.creatures.IAttackable;
 import com.group4.app.model.creatures.ICanAttack;
 import com.group4.app.model.creatures.Player;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.dungeon.Tile;
 import com.group4.app.model.dungeon.World;
 import com.group4.app.model.items.WeaponFactory;
