+++ src/test/java/com/group4/app/testmodel/TestCreature.java
@@ -3,6 +3,7 @@ package com.group4.app.testmodel;
 import com.group4.app.model.*;
 import com.group4.app.model.creatures.Attributes;
 import com.group4.app.model.creatures.Creature;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.items.ArmourFactory;
 import com.group4.app.model.items.ArmourType;
 import com.group4.app.model.items.Weapon;
