+++ src/test/java/com/group4/app/testmodel/TestEnemy.java
@@ -1,12 +1,11 @@
 package com.group4.app.testmodel;
 
 import com.group4.app.model.*;
-import com.group4.app.model.Model;
-import com.group4.app.model.Position;
 import com.group4.app.model.creatures.Attributes;
 import com.group4.app.model.creatures.Enemy;
 import com.group4.app.model.creatures.EnemyFactory;
 import com.group4.app.model.creatures.Player;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.dungeon.Tile;
 import com.group4.app.model.dungeon.World;
 import com.group4.app.model.items.WeaponFactory;
