+++ src/test/java/com/group4/app/testmodel/TestTurnHandler.java
@@ -7,9 +7,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import org.junit.jupiter.api.Test;
 
-import com.group4.app.model.ITurnTaker;
-import com.group4.app.model.TurnHandler;
 import com.group4.app.model.creatures.EnemyFactory;
+import com.group4.app.model.turns.ITurnTaker;
+import com.group4.app.model.turns.TurnHandler;
 
 public class TestTurnHandler {
