+++ src/test/java/com/group4/app/testmodel/TestWorld.java
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import org.junit.jupiter.api.Test;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.Position;
 import com.group4.app.model.dungeon.Tile;
 import com.group4.app.model.dungeon.World;
